feat!: remove url token auth
