+++ .env
@@ -1,4 +1,3 @@
-URL_ACCESS_TOKEN=test_token
 API_BASE_URL=http://localhost:3000
 DEBUG=true
