+++ tests/.env.base
@@ -1,5 +1,4 @@
-URL_ACCESS_TOKEN=test_token
 API_BASE_URL=http://localhost:3000
 NODE_ENV=development
 RUNNER_TEMP=tests/temp/runner
-ACTIONS_CACHE_URL=http://localhost:3000/test_token/
+ACTIONS_CACHE_URL=http://localhost:3000/
